Move undefined to models
