+++ src/models/mod.rs
@@ -2,8 +2,10 @@ mod apis;
 mod http;
 mod keys;
 mod ratelimit;
+mod undefined;
 
 pub use apis::*;
 pub use http::*;
 pub use keys::*;
 pub use ratelimit::*;
+pub use undefined::*;
